requirements.txt
Django==1.6.1
celery==3.1.9
django-celery==3.1.9
django-extensions==1.3.2
pysqlite==2.6.3
pytz==2013.9
django-widget-tweaks==1.3
django-tastypie==0.12.1
django-postman<3.3
django-solo<1.2
werkzeug<0.11